feat(storage-provider): Move benchmarking and associated modules into a directory
